backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt
django-cors-headers
django-filter
requests>=2.31.0
